move helpers/nodeId.ts -> src/identifiers.ts
