+++ packages/dht/src/connection/ConnectionLockRpcLocal.ts
@@ -17,7 +17,7 @@ import { IConnectionLockRpc } from '../proto/packages/dht/protos/DhtRpc.server'
 import { DhtCallContext } from '../rpc-protocol/DhtCallContext'
 import { getNodeIdOrUnknownFromPeerDescriptor } from './ConnectionManager'
 import { LockID } from './ConnectionLockHandler'
-import { NodeID } from '../helpers/nodeId'
+import { NodeID } from '../identifiers'
 
 interface ConnectionLockRpcLocalConfig {
     addRemoteLocked: (id: NodeID, lockId: LockID) => void
