+++ packages/dht/src/helpers/PeerID.ts
@@ -2,7 +2,7 @@ import { BrandedString, binaryToHex } from '@streamr/utils'
 import { UUID } from './UUID'
 import { IllegalArguments } from './errors'
 import crypto from 'crypto'
-import { NodeID, getNodeIdFromBinary } from './nodeId'
+import { NodeID, getNodeIdFromBinary } from '../identifiers'
 
 export type PeerIDKey = BrandedString<'PeerIDKey'>
