+++ packages/dht/test/unit/RandomContactList.test.ts
@@ -1,6 +1,6 @@
 import { RandomContactList } from '../../src/dht/contact/RandomContactList'
 import { PeerID } from '../../src/helpers/PeerID'
-import { NodeID, getNodeIdFromBinary } from '../../src/helpers/nodeId'
+import { NodeID, getNodeIdFromBinary } from '../../src/identifiers'
 
 const createItem = (nodeId: Uint8Array): { getNodeId: () => NodeID, getPeerId: () => PeerID } => {
     return {
